+++ src/Contracts/Repositories/UserWriteRepository.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace TJVB\GitlabModelsForLaravel\Contracts\Repositories;
+
+interface UserWriteRepository
+{
+    public function updateOrCreate(int $userId, array $userdata): ?int;
+}